requirements.txt
# Web framework
flask==3.0.3
flask-login==0.6.3      # easy session-based authentication
flask-bcrypt==1.0.1     # password hashing

# Borg orchestration
borgbackup==1.4.1
msgpack==1.1.0
python-dotenv==1.0.1    # env var config support

# Optional dev utilities
sqlalchemy==2.0.32      # ORM for SQLite (recommended over raw sqlite3)
flask_sqlalchemy        # Flask integration for SQLAlchemy
